+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * added the `ADMIN_PREFIX` constant;
 * updated for MeTools 2.11.1;
 * added tests for `AuthHelper`, `MenuHelper` and `Plugin` classes;
-* added tests for global functions.
+* added tests for global functions and request detectors.
 
 ### 2.14.10
 * updated for MeTools 2.11.0.